Move data splitters into preprocessing
