+++ evalml/automl/__init__.py
@@ -1,3 +1,2 @@
 from .automl_search import AutoMLSearch
 from .utils import get_default_primary_search_objective, make_data_splitter
-from .data_splitters import TrainingValidationSplit, TimeSeriesSplit
