+++ evalml/preprocessing/__init__.py
@@ -5,3 +5,4 @@ from .utils import (
     target_distribution,
     drop_nan_target_rows
 )
+from .data_splitters import TrainingValidationSplit, TimeSeriesSplit
